+++ src/adapter/admin/groups.mjs
@@ -60,6 +60,10 @@ class Group {
     return snap.data().result;
   }
 
+  as() {
+    return this;
+  }
+
   adapter() {
     return {
       db: () => this.db,