+++ src/tests/get.ts
@@ -351,12 +351,14 @@ describe("get", () => {
     it("exposes the request", () => {
       const id = db.users.id("whatever");
       const promise = db.users.get(id);
-      expect(promise.request).toEqual({
-        type: "request",
-        kind: "get",
-        path: "users",
-        id,
-      });
+      expect(promise.request).toEqual(
+        expect.objectContaining({
+          type: "request",
+          kind: "get",
+          path: "users",
+          id,
+        }),
+      );
     });
   });
 });