+++ src/tests/query.ts
@@ -2242,39 +2242,43 @@ describe("query", () => {
   describe("request", () => {
     it("exposes the request", () => {
       const promise = db.contacts.query(($) => $.field("ownerId").eq(ownerId));
-      expect(promise.request).toEqual({
-        type: "request",
-        kind: "query",
-        path: "contacts",
-        queries: [
-          {
-            field: ["ownerId"],
-            filter: "==",
-            type: "where",
-            value: ownerId,
-          },
-        ],
-      });
+      expect(promise.request).toEqual(
+        expect.objectContaining({
+          type: "request",
+          kind: "query",
+          path: "contacts",
+          queries: [
+            {
+              field: ["ownerId"],
+              filter: "==",
+              type: "where",
+              value: ownerId,
+            },
+          ],
+        }),
+      );
     });
 
     it("exposes a group request", () => {
       const promise = groups(db).contacts.query(($) =>
         $.field("ownerId").eq(ownerId),
       );
-      expect(promise.request).toEqual({
-        type: "request",
-        kind: "query",
-        path: "contacts",
-        queries: [
-          {
-            field: ["ownerId"],
-            filter: "==",
-            type: "where",
-            value: ownerId,
-          },
-        ],
-        group: true,
-      });
+      expect(promise.request).toEqual(
+        expect.objectContaining({
+          type: "request",
+          kind: "query",
+          path: "contacts",
+          queries: [
+            {
+              field: ["ownerId"],
+              filter: "==",
+              type: "where",
+              value: ownerId,
+            },
+          ],
+          group: true,
+        }),
+      );
     });
   });
 });